+++ persist.js
@@ -95,7 +95,7 @@
 	}
 
 	Molpy.Flood = function() {
-		return moment.utc([13291, 4, 10, 0, 0, 0, 0]).format();
+		return dayjs.utc([13291, 4, 10, 0, 0, 0, 0]).format();
 	}
 
 	Molpy.Load = function() {
@@ -313,8 +313,9 @@
 		          + (Molpy.Redacted.chainCurrent) + s
 		          + (Molpy.Redacted.chainMax) + s
 		          + (Molpy.lootPerPage) + s
+		          + (Molpy.time.toDate().getTime()) + s
 		          + (Molpy.largestNPvisited[0]) + s;
-		          for(var i=0;i<Molpy.fracParts.length;i++){str=str+ (Molpy.largestNPvisited[Molpy.fracParts[i]]) + s}
+        for(var i=0;i<Molpy.fracParts.length;i++){str=str+ (Molpy.largestNPvisited[Molpy.fracParts[i]]) + s}
 		return str;
 	}
 
@@ -555,9 +556,14 @@
 			Molpy.Redacted.chainCurrent = parseFloat(pixels[13]) || 0;
 			Molpy.Redacted.chainMax = parseFloat(pixels[14]) || 0;
 			Molpy.lootPerPage = parseInt(pixels[15]) || 20;
-			Molpy.largestNPvisited[0] = (parseInt(pixels[16]) || parseFloat(pixels[16]))||Math.abs(Molpy.newpixNumber);
+            var e = 0;
+            if(version>=4.1){
+                Molpy.time=dayjs(parseInt(pixels[16]))
+                e=1;
+            }
+			Molpy.largestNPvisited[0] = (parseInt(pixels[16+e]) || parseFloat(pixels[16+e]))||Math.abs(Molpy.newpixNumber);
 			for(var i=0;i<Molpy.fracParts.length;i++){
-				Molpy.largestNPvisited[Molpy.fracParts[i]]=parseFloat(pixels[17+i])||0}
+				Molpy.largestNPvisited[Molpy.fracParts[i]]=parseFloat(pixels[17+e+i])||0}
 		}
 	};
 
@@ -928,7 +934,13 @@
 		Molpy.MustardCheck();
 		Molpy.RefreshOptions();
 
-		Molpy.ONGstart = ONGsnip(moment()); //if you missed the ONG before loading, too bad!
+		Molpy.ONGstart = ONGsnip(dayjs()); //if you missed the ONG before loading, too bad!
+        if(version >= 4.1 && Molpy.IsEnabled('PoG')){
+            var oldStart = ONGsnip(Molpy.time);
+            var diff = (Molpy.ONGstart - oldStart)/1000;
+            var np = diff/Molpy.NPlength;
+            Molpy.Add('Shork',np);
+        }
 		g('clockface').className = Molpy.Boosts['Coma Molpy Style'].power ? 'hidden' : 'unhidden';
 		Molpy.HandlePeriods();
 		Molpy.UpdateBeach();
@@ -1258,9 +1270,9 @@
 			Molpy.BoostsOwned = 0;
 			Molpy.notifsReceived = 0;
 
-			Molpy.startDate = parseInt(moment().valueOf());
+			Molpy.startDate = parseInt(dayjs().valueOf());
 			Molpy.newpixNumber = 1;
-			Molpy.ONGstart = ONGsnip(moment());
+			Molpy.ONGstart = ONGsnip(dayjs());
 			Molpy.DefaultOptions();
 
 			var keep = '';